+++ CHANGELOG.md
@@ -17,7 +17,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 <!-- ### Fixed -->
 <!-- ### Removed -->
 
-<!-- ## Unreleased -->
+## Unreleased
+
+### Fixed
+
+- TypeScript errors of the form `Property '...' does not exist on type '...'`
+  are now suppressed (temporarily until `d.ts` files are fetched).
 
 ## [0.6.6] - 2021-03-29