+++ CHANGELOG.md
@@ -26,6 +26,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   type as the JSON config, and can be used to save/restore the state of all
   files and other configuration.
 
+### Fixed
+
+- TypeScript errors of the form `Property '...' does not exist on type '...'`
+  are now suppressed (temporarily until `d.ts` files are fetched).
+
 ## [0.6.6] - 2021-03-29
 
 ### Fixed